Unify CameraEvents into wrapper class
