--- glass/WearScript/src/main/java/com/dappervision/wearscript/JSEvents/CameraEvent.java
@@ -1,11 +1,0 @@
-package com.dappervision.wearscript.jsevents;
-
-public class CameraEvent implements JSBusEvent {
-    private double period;
-    public CameraEvent(double period){
-        this.period = period;
-    }
-    public double getPeriod(){
-        return period;
-    }
-}
