--- glass/WearScript/src/main/java/com/dappervision/wearscript/JSEvents/CameraPhotoEvent.java
@@ -1,7 +1,0 @@
-package com.dappervision.wearscript.jsevents;
-
-public class CameraPhotoEvent extends CallbackJSBusEvent {
-    public CameraPhotoEvent(String callback) {
-        super(callback);
-    }
-}
