--- glass/WearScript/src/main/java/com/dappervision/wearscript/JSEvents/CameraVideoEvent.java
@@ -1,7 +1,0 @@
-package com.dappervision.wearscript.jsevents;
-
-public class CameraVideoEvent extends CallbackJSBusEvent {
-    public CameraVideoEvent(String callback) {
-        super(callback);
-    }
-}
