+++ glass/WearScript/src/main/java/com/dappervision/wearscript/WearScript.java
@@ -1,12 +1,16 @@
 package com.dappervision.wearscript;
 
+import com.dappervision.wearscript.activities.MainActivity;
+import com.dappervision.wearscript.dataproviders.DataPoint;
+import com.dappervision.wearscript.events.LogEvent;
+import com.dappervision.wearscript.events.SendBlobEvent;
+import com.dappervision.wearscript.events.ServerConnectEvent;
+import com.dappervision.wearscript.events.ShutdownEvent;
 import com.dappervision.wearscript.jsevents.ActivityEvent;
 import com.dappervision.wearscript.jsevents.BarcodeCallbackEvent;
 import com.dappervision.wearscript.jsevents.BlobCallbackEvent;
 import com.dappervision.wearscript.jsevents.CameraCallbackEvent;
-import com.dappervision.wearscript.jsevents.CameraEvent;
-import com.dappervision.wearscript.jsevents.CameraPhotoEvent;
-import com.dappervision.wearscript.jsevents.CameraVideoEvent;
+import com.dappervision.wearscript.jsevents.CameraEvents;
 import com.dappervision.wearscript.jsevents.DataLogEvent;
 import com.dappervision.wearscript.jsevents.GestureCallbackEvent;
 import com.dappervision.wearscript.jsevents.LiveCardEvent;
@@ -19,12 +23,6 @@ import com.dappervision.wearscript.jsevents.SpeechRecognizeEvent;
 import com.dappervision.wearscript.jsevents.WifiCallbackEvent;
 import com.dappervision.wearscript.jsevents.WifiEvent;
 import com.dappervision.wearscript.jsevents.WifiScanEvent;
-import com.dappervision.wearscript.activities.MainActivity;
-import com.dappervision.wearscript.dataproviders.DataPoint;
-import com.dappervision.wearscript.events.LogEvent;
-import com.dappervision.wearscript.events.SendBlobEvent;
-import com.dappervision.wearscript.events.ServerConnectEvent;
-import com.dappervision.wearscript.events.ShutdownEvent;
 
 import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
@@ -131,23 +129,23 @@ public class WearScript {
     }
 
     public void cameraOff() {
-        getEventBus().post(new CameraEvent(0));
+        getEventBus().post(new CameraEvents.Start(0));
     }
 
     public void cameraPhoto() {
-        getEventBus().post(new CameraPhotoEvent(null));
+        getEventBus().post(new CameraEvents.Photo(null));
     }
 
     public void cameraPhoto(String callback) {
-        getEventBus().post(new CameraPhotoEvent(callback));
+        getEventBus().post(new CameraEvents.Photo(callback));
     }
 
     public void cameraVideo() {
-        getEventBus().post(new CameraVideoEvent(null));
+        getEventBus().post(new CameraEvents.Video(null));
     }
 
     public void cameraOn(double imagePeriod) {
-        getEventBus().post(new CameraEvent(imagePeriod));
+        getEventBus().post(new CameraEvents.Start(imagePeriod));
     }
 
     public void cameraCallback(int type, String callback) {
